app/requirements.txt
fastapi==0.111.0
pydantic==2.8.2
pandas==2.2.2
python-dotenv==0.21.0
python-multipart==0.0.9
uvicorn==0.20.0
imbalanced-learn==0.12.3
gunicorn==22.0.0
pydantic-settings==2.3.4
scikit-learn==1.5.0